requirements.txt
Flask==2.0.1
Flask-SocketIO==5.3.6
opencv-contrib-python==4.5.5.64
numpy~=1.26.4
Pillow~=11.1.0
pytest==8.3.2
pytest-html==4.1.1
matplotlib~=3.10.1
pandas==2.2.2
mediapipe~=0.10.21
face-recognition~=1.3.0